chore: Reorganize some files.
